fixed some issue and moved PlaylistFileType to own file
